골드 text ui 추가 , 골드 획득 로직 추가
골드 text ui 추가 , 골드 획득 로직 추가

+++ Assets/WOM/Scripts/Data/DataModel.cs
@@ -98,7 +98,7 @@ public class MonsterData
     public int monsterId;
     public float hp;
     public float exp;
-    public float gold;
+    public int gold;
     public int goldCount;
     public int imageId;
     public int bgId;
