+++ Assets/WOM/Scripts/Data/MonsterBase.cs
@@ -8,7 +8,7 @@ public class MonsterBase : MonoBehaviour
     public int monsterId;
     public float hp;
     public float exp;
-    public float gold;
+    public int gold;
     public int goldCount;
     public int imageId;
     public int bgId;
